// src/app/shared/models/address.model.ts
export interface AddressModel {
  'addressName'?: string;
  'city'?: string;
  'contact'?: string;
  'contact2'?: string;
  'country'?: string;
  'dic'?: string;
  'email'?: string;
  'ico'?: string;
  'phone'?: string;
  'resource'?: 'SYSTEM';
  'resourceId'?: string;
  'street'?: string;
  'type'?: 'DELIVERY';
  'zip'?: string;
}